% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{compare_proportions}
\alias{compare_proportions}
\title{Compare Proportions (Simple)}
\usage{
compare_proportions(data, conf.level = 0.95, method = "holm")
}
\arguments{
\item{data}{Data frame with 'proportion' and 'trials' columns}

\item{conf.level}{Confidence level}

\item{method}{Adjustment method}
}
\description{
Pairwise comparison of proportions using normal approximation
}
\examples{
data <- data.frame(proportion = c(0.3, 0.5), trials = c(50, 50), group = c("A", "B"))
compare_proportions(data)

}
