% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_tibble}
\alias{format_tibble}
\title{Format Tibble for Display}
\usage{
format_tibble(data, digits = 2)
}
\arguments{
\item{data}{Tibble or data frame}

\item{digits}{Number of digits to round to (default: 2)}
}
\value{
Formatted tibble
}
\description{
Formats numeric columns in a tibble for display with specified digits
}
