% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{kk_pal}
\alias{kk_pal}
\title{KK Colour Palette}
\usage{
kk_pal(n, colors = getOption("kkstatfun.colors"))
}
\arguments{
\item{n}{Number of colours required.}

\item{colors}{Character vector of anchor colours (hex). Defaults to the palette
registered by \code{\link[=set_plot_colors]{set_plot_colors()}}, or a red/navy/amber flag if none is set.}
}
\value{
Character vector of \code{n} hex colours.
}
\description{
Generates \code{n} colours from a set of anchor ("flag") colours. When
\code{n} does not exceed the number of anchors the anchor colours are used
directly; when more are needed the anchors are interpolated with
\code{\link[grDevices:colorRampPalette]{grDevices::colorRampPalette()}}. The anchors default to those set by
\code{\link[=set_plot_colors]{set_plot_colors()}}.
}
\examples{
kk_pal(2)
kk_pal(6, colors = c("#D62828", "#003049", "#F77F00"))

}
