% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kk_setup}
\alias{kk_setup}
\title{Initialize KK Stat Fun Environment}
\usage{
kk_setup(cores = parallel::detectCores(logical = FALSE) - 1, scipen = 999)
}
\arguments{
\item{cores}{Number of cores to use (default: detectCores() - 1)}

\item{scipen}{Scientific notation penalty (default: 999)}
}
\description{
Sets preferred options for backend, parallel processing, and display.
}
