% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{power_proportions}
\alias{power_proportions}
\title{Power Analysis for Proportions}
\usage{
power_proportions(
  n = NULL,
  p1 = NULL,
  p2 = NULL,
  sig.level = 0.05,
  power = NULL,
  alternative = "two.sided",
  strict = FALSE
)
}
\arguments{
\item{n}{Number of observations (per group)}

\item{p1}{Proportion in group 1}

\item{p2}{Proportion in group 2}

\item{sig.level}{Significance level (Type I error probability)}

\item{power}{Power of test (1 minus Type II error probability)}

\item{alternative}{One- or two-sided test ("two.sided", "one.sided")}

\item{strict}{Use strict interpretation in two-sided case}
}
\value{
Object of class "power.htest"
}
\description{
Calculate the necessary sample size or power for detecting a difference in proportions.
This is a wrapper around \code{stats::power.prop.test}.
}
