% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{prop_trend_test}
\alias{prop_trend_test}
\title{Test for Trend in Proportions}
\usage{
prop_trend_test(data, x, n = NULL, group = NULL)
}
\arguments{
\item{data}{Data frame}

\item{x}{Variable for number of successes (or binary outcome if n is NULL)}

\item{n}{Variable for number of trials (optional if x is binary outcome)}

\item{group}{Grouping variable (ordered)}
}
\value{
Result of prop.trend.test
}
\description{
Performs the Cochran-Armitage test for trend in proportions.
}
\examples{
data <- data.frame(
              dose = c("Low", "Medium", "High"),
              positive = c(10, 20, 30),
              total = c(100, 100, 100)
)
prop_trend_test(data, positive, total, dose)

}
