% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival_plot}
\alias{survival_plot}
\title{Survival Analysis Plot}
\usage{
survival_plot(data, time_col, status_col, group_col = NULL, conf.int = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{time_col}{Time variable name}

\item{status_col}{Status variable name (0=censored, 1=event)}

\item{group_col}{Optional grouping variable}

\item{conf.int}{Show confidence intervals}
}
\value{
ggplot object
}
\description{
Creates a Kaplan-Meier survival plot with risk table
}
\examples{
library(survival)
survival_plot(lung, "time", "status", "sex")

}
