% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.R
\name{compare_proportions_by}
\alias{compare_proportions_by}
\title{Compare Proportions Stratified by Group}
\usage{
compare_proportions_by(data, conf.level = 0.95, method = "holm")
}
\arguments{
\item{data}{Data frame}

\item{conf.level}{Confidence level}

\item{method}{Adjustment method}
}
\description{
Compares proportions within groups (stratified analysis)
}
\examples{
data <- data.frame(
              group = rep(c("A", "B"), each = 2),
              subgroup = rep(c("X", "Y"), 2),
              proportion = c(0.3, 0.5, 0.4, 0.6),
              trials = rep(50, 4)
)
compare_proportions_by(data)

}
