% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egn.R
\name{extract_age_from_egn}
\alias{extract_age_from_egn}
\title{Extract Information from EGN}
\usage{
extract_age_from_egn(egn, admission_date = Sys.Date())
}
\arguments{
\item{egn}{Vector of EGNs (character or numeric)}

\item{admission_date}{Date to calculate age against (default: Sys.Date())}
}
\value{
Data frame with extracted information
}
\description{
Parses Bulgarian EGN to extract age, birth date, gender, and region.
Also validates the EGN.
}
\examples{
extract_age_from_egn("7501010010")

}
