% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bland_altman.R
\name{kk_bland_altman}
\alias{kk_bland_altman}
\title{Bland-Altman Plot and Analysis (KK)}
\usage{
kk_bland_altman(
  method1,
  method2,
  conf.level = 0.95,
  plot = TRUE,
  title = "Bland-Altman Plot"
)
}
\arguments{
\item{method1}{Vector of measurements from method 1}

\item{method2}{Vector of measurements from method 2}

\item{conf.level}{Confidence level for limits of agreement (default 0.95)}

\item{plot}{Return a ggplot object? (default TRUE)}

\item{title}{Plot title}
}
\value{
If plot=TRUE, a ggplot object. If plot=FALSE, a tibble with stats.
}
\description{
Performs Bland-Altman analysis to assess agreement between two methods of measurement.
}
\examples{
m1 <- c(10, 12, 15, 20, 25)
m2 <- c(11, 13, 14, 21, 24)
kk_bland_altman(m1, m2)
}
