% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_stats.R
\name{kk_risk_plot}
\alias{kk_risk_plot}
\title{Risk Plot (KK)}
\usage{
kk_risk_plot(data, title = "Risk Estimates")
}
\arguments{
\item{data}{Data frame containing Metric, Estimate, Lower, Upper (output of kk_epi_2x2)}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Forest plot for Odds Ratios or Relative Risks.
}
\examples{
res <- tibble::tibble(
              Metric = c("OR", "RR"),
              Estimate = c(1.5, 1.2),
              Lower = c(1.1, 1.0),
              Upper = c(2.0, 1.5)
)
kk_risk_plot(res)

}
