% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{kk_table1}
\alias{kk_table1}
\title{Wrapper for Table 1 Generation}
\usage{
kk_table1(
  data,
  by = NULL,
  variables = NULL,
  label_list = NULL,
  digits = 2,
  p_value = TRUE
)
}
\description{
A wrapper around \code{table1_summary} to generate a publication-ready descriptive table.
}
\examples{
\dontrun{
kk_table1(mtcars, by = "am", variables = c("mpg", "hp"))
}
}
