% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{kkonehot}
\alias{kkonehot}
\title{One-Hot Encoding (KK One Hot)}
\usage{
kkonehot(data, column)
}
\arguments{
\item{data}{Tibble or data frame}

\item{column}{Column name (character)}
}
\value{
Data with new binary columns added
}
\description{
Creates binary dummy variables for categorical column
}
