% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{kkplot}
\alias{kkplot}
\title{KK Plot Wrapper}
\usage{
kkplot(...)
}
\arguments{
\item{...}{Arguments passed to ggplot}
}
\description{
Wrapper for ggplot with axis caps
}
\examples{
kkplot(mtcars, aes(x = mpg, y = wt)) + geom_point()

}
