% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_round}
\alias{mutate_round}
\title{Round Numeric Columns}
\usage{
mutate_round(data, digits = 2)
}
\arguments{
\item{data}{Tibble or data frame}

\item{digits}{Number of digits (default: 2)}
}
\value{
Data frame with rounded numeric columns
}
\description{
Rounds all numeric columns in a data frame
}
\examples{
df <- data.frame(a = c(1.234, 5.678), b = c("x", "y"))
mutate_round(df, 1)

}
