% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_proportion_comparisons}
\alias{plot_proportion_comparisons}
\title{Plot Proportion Comparisons}
\usage{
plot_proportion_comparisons(
  results,
  title = "Proportion Comparisons",
  subtitle = NULL,
  xlab = "Difference in Proportions (95\% CI)"
)
}
\arguments{
\item{results}{Data frame returned by \code{compare_proportions} or \code{compare_proportions_kk_glm}}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{xlab}{Label for x-axis}
}
\value{
ggplot object
}
\description{
Visualizes the output of \code{compare_proportions} or \code{compare_proportions_kk_glm} using a forest plot.
}
