% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1_summary}
\alias{table1_summary}
\title{Create Publication-Ready Table 1}
\usage{
table1_summary(data, by = NULL, variables = NULL, label_list = NULL)
}
\arguments{
\item{data}{Data frame}

\item{by}{Grouping variable (character)}

\item{variables}{Vector of variables to include (optional, defaults to all)}

\item{label_list}{Named list of labels for variables}
}
\value{
gtsummary object
}
\description{
Generates a summary table of descriptive statistics, stratified by a group
variable, with p-values.
}
\examples{
\dontrun{
library(gtsummary)
table1_summary(mtcars, by = "am", variables = c("mpg", "hp", "wt"))
}

}
