% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{univariate_cat_plot}
\alias{univariate_cat_plot}
\title{Univariate Categorical Plot}
\usage{
univariate_cat_plot(data, variable)
}
\arguments{
\item{data}{Data frame}

\item{variable}{Variable to plot}
}
\description{
Univariate Categorical Plot
}
