% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{univariate_cont_plot}
\alias{univariate_cont_plot}
\alias{univariate_continuous_plot}
\title{Univariate Continuous Plot}
\usage{
univariate_cont_plot(
  data,
  variable,
  group = NULL,
  label_size = 3.5,
  stats = "mean"
)

univariate_continuous_plot(
  data,
  variable,
  group = NULL,
  label_size = 3.5,
  stats = "mean"
)
}
\arguments{
\item{data}{Data frame}

\item{variable}{Variable to plot}

\item{group}{Optional grouping variable}

\item{label_size}{Size for text labels (default: 3.5)}
}
\description{
Univariate Continuous Plot
}
\examples{
univariate_cont_plot(mtcars, "mpg")
univariate_cont_plot(mtcars, "mpg", group = "cyl")

}
