% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{univariate_cont_plot}
\alias{univariate_cont_plot}
\title{Univariate Continuous Plot}
\usage{
univariate_cont_plot(data, variable)
}
\arguments{
\item{data}{Data frame}

\item{variable}{Variable to plot}
}
\description{
Univariate Continuous Plot
}
