% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{univariate_plot}
\alias{univariate_plot}
\title{Univariate Plot}
\usage{
univariate_plot(data, variable)
}
\arguments{
\item{data}{Data frame}

\item{variable}{Variable to plot}
}
\description{
Automatically detects variable type and plots accordingly
}
\examples{
univariate_plot(mtcars, "mpg")
univariate_plot(mtcars, "am")

}
